test! test! test!

+++ lib/std/os.go
@@ -1,11 +1,11 @@
-func exit_with(status int) noreturn {
-  asm {
-    "movq $60, %rax",
-    "syscall"
-  };
+func exit_with(status int64) noreturn {
+    asm {
+        "movq $60, %rax",
+        "syscall"
+    };
 }
 
-func write(fd int, buf str, count int) noreturn {
+func write(fd int64, buf str, count int64) noreturn {
     asm {
         "movq $1, %rax",
         "syscall"
