+++ lib/std/version.go
@@ -1,3 +1,3 @@
-func peachili_version() int {
+func peachili_version() int64 {
     return 1;
 }
